Create a converter script for database files created by the old Markov bot
The old bot used pickle to serialize its data files, so I had to write this script in python in order to be able to read that format. It wasn't super desirable but it had to be done.

To use the converter script, simply run it with python and pass the path to the .dat file you would like to convert as an argument.