Remove unicode category restrictions
This stuff was in the original markov bot to prevent weird characters from entering the data files. I just discovered that it blocks emojis, which I think detracts from the bot experience a bit. I think that the problem of having weird characters like RTL or something like that in the data isn't really that big of a problem, and if its too big of an issue in some chat the administrator can just clear the chain.